+++ mod/attendance/classes/calendar_helpers.php
@@ -132,11 +132,12 @@ function attendance_update_calendar_event($session) {
         return attendance_create_calendar_event($session);
     }
 
-    // Boring update
+    // Boring update.
     $caleventdata = new stdClass();
     $caleventdata->timeduration   = $timeduration;
     $caleventdata->timestart      = $timestart;
     $caleventdata->timemodified   = time();
+    $caleventdata->description    = $session->description;
 
     $calendarevent = calendar_event::load($caleventid);
     if ($calendarevent) {