+++ mod/bigbluebuttonbn/db/access.php
@@ -66,20 +66,22 @@ $capabilities = array(
 
     // ecastro ULPGC capabilities
     // Ability to generate live meeting.
+    //'mod/bigbluebuttonbn:meeting' => array(
     'mod/bigbluebuttonbn:room' => array(
         'captype' => 'write',
         'contextlevel' => CONTEXT_MODULE,
         'archetypes' => array(
             'manager' => CAP_ALLOW,
             'editingteacher' => CAP_ALLOW,
-            'teacher' => CAP_PREVENT,
-            'student' => CAP_PREVENT,
+            'teacher' => CAP_ALLOW,
+            'student' => CAP_ALLOW,
             'guest' => CAP_PREVENT,
         ),
     ),
     
     
     // Ability to record a meeting.
+    //'mod/bigbluebuttonbn:recording' => array(
     'mod/bigbluebuttonbn:record' => array(
         'captype' => 'write',
         'contextlevel' => CONTEXT_MODULE,